fix: Move server list syncing to TextChannel to fix unable to @ members
